requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiogram==2.4
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.2
annotated-types==0.6.0
argcomplete==3.2.1
args==0.1.0
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.28.0
Babel==2.14.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
bottle==0.12.23
bottle-websocket==0.2.9
bpy==3.5.0
browser-cookie3==0.19.1
cachetools==5.3.0
certifi==2023.11.17
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
clint==0.5.1
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.6
contourpy==1.0.6
cryptography==41.0.7
curl-cffi==0.5.5
customtkinter==5.2.1
cycler==0.11.0
Cython==0.29.35
darkdetect==0.8.0
dearpygui==1.9.1
decorator==5.1.1
distlib==0.3.8
Django==4.1.5
docutils==0.19
docx==0.2.4
docx2txt==0.8
easygui==0.98.3
ebcdic==1.1.1
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
extract-msg==0.47.0
fake-useragent==1.1.3
fastrlock==0.8.1
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.1
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.12.2
future==0.18.2
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.51.3
h11==0.9.0
h2==3.2.0
h3==3.7.6
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.19.4
hyperframe==5.2.0
idna==2.10
image==1.5.33
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
keyboard==0.13.5
keyring==23.13.1
kiwisolver==1.4.4
lark==1.1.8
Levenshtein==0.21.1
libclang==16.0.0
libusb==1.0.26b5
libusb1==3.0.0
llvmlite==0.41.1
lxml==4.9.3
lz4==4.3.2
magic-filter==1.0.12
MailGw-Temporary-Email==0.0.2
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
more-itertools==9.0.0
MouseInfo==0.1.3
mpmath==1.3.0
msoffcrypto-tool==5.1.1
mss==7.0.1
multidict==6.0.4
names==0.3.0
networkx==3.2.1
nominatim==0.1
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
olefile==0.47
oletools==0.60.1
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.7.0.68
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
pandas==1.5.3
path==16.9.0
pcodedmp==1.2.6
pefile==2022.5.30
Pillow==9.3.0
pim==1.0.1
pkg-about==1.0.8
pkginfo==1.9.6
platformdirs==4.1.0
protobuf==3.19.6
provide-dir==0.1.2
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
PyExecJS==1.5.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pymailtm==1.1.1
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPasser==0.0.5
pyperclip==1.8.2
PyQt5==5.15.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5-stubs==5.15.6.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pyth==0.6.0
pyth3==0.7
python-dateutil==2.8.2
python-docx==1.1.0
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
pywin32-ctypes==0.2.0
pyxel==1.9.16
PyYAML==6.0.1
random-password-generator==2.2.0
random-username==1.0.2
rapidfuzz==3.1.1
readme-renderer==37.3
red-black-tree-mod==1.20
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
RTFDE==0.1.1
rtfparse==0.8.1
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.10.0
selenium==4.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
sqlparse==0.4.3
srt==3.5.3
streamlit==1.21.0
streamlit-chat @ https://github.com/AI-Yash/st-chat/archive/refs/pull/24/head.zip#sha256=c4e9309dfb2d0d7859c9d26051e62c1ba3d7f352ceb08bb950e1a3f6645bbfc8
striprtf==0.0.26
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tiktoken==0.5.2
timezonefinder==6.1.9
tls-client==0.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.36.1
trio==0.22.0
trio-websocket==0.10.2
twine==4.0.2
TwoCaptcha==0.0.1
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.14
usb==0.0.83.dev0
validators==0.20.0
virtualenv==20.25.0
vosk==0.3.45
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==2.3.6
whichcraft==0.6.1
win-unicode-console==0.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
zstandard==0.21.0
